Create README.md
overview
A fast moving consumer goods company entered into the instant noodles business two years back.  The higher management noticed that there is a mismatch in the demand and supply. Where the demand is high, supply is pretty low and vice versa which results in a loss in inventory cost and ultimately lost to the company. Hence the higher management wants to optimize the supply quantity in each and every warehouse in the entire country.